requirements.txt
flask
python-dotenv
groq
requests
pandas